backend/requirements.txt
google-generativeai
pymongo
pytest
fastapi
uvicorn
motor
python-dotenv